Fix third_party/gpgmm GN build.
Removes a circular dependency on TraceEvent between source and common targets by moving the source dependant code out of common.
